Fix: Background worker not starting on replicas
The background worker that's responsible for log rotation doesn't start
on the replica nodes, which results in all audit messages going to a
single file until the server is restarted.

According to the PostgreSQL documentation, ConsistentState and
RecoveryFinished are equivalent on the leader, and ConsistentState is
enough on the hot replicas, meaning it should work for both node roles.